Commit dos metodos da venda

+++ pdv-api/src/main/java/org/example/pdvapi/repositories/ItemVendaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemVendaRepository extends JpaRepository<ItemVenda, Long> {
+public interface ItemVendaRepository extends JpaRepository<ItemVenda, Integer> {
 }
